refactor: simplify HomePage component by removing mouse event handlers and integrating InteractiveCard for improved code organization
